Clean up ConnectionManager
